requirements.txt
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
attrs==24.3.0
billiard==4.2.1
black==25.1.0
celery==5.5.2
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coverage==7.6.10
cron-descriptor==1.4.5
cryptography==44.0.0
defusedxml==0.8.0rc2
distro==1.9.0
Django==5.1.5
django-celery-beat==2.8.0
django-ckeditor-5==0.2.15
django-debug-toolbar==5.0.1
django-filter==24.3
django-grappelli==4.0.2
django-js-asset==3.0.1
django-mptt==0.16.0
django-object-actions==5.0.0
django-timezone-field==7.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-spectacular==0.28.0
environs==14.1.0
factory_boy==3.3.1
Faker==35.0.0
flake8==7.2.0
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
isort==6.0.1
jiter==0.10.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.3
Markdown==3.7
marshmallow==3.25.1
mccabe==0.7.0
mypy_extensions==1.1.0
oauthlib==3.2.2
openai==1.82.1
packaging==24.2
pathspec==0.12.1
pillow==11.1.0
platformdirs==4.3.8
pluggy==1.5.0
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.5
pydantic_core==2.33.2
pyflakes==3.3.2
PyJWT==2.10.1
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.3
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.12.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
